api/workspaces: when merging pull request via sturdy, do not wait for the webhook
Created with Sturdy
