+++ api/pkg/github/enterprise/service/pull_request.go
@@ -15,17 +15,21 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/oauth2"
 
+	"getsturdy.com/api/pkg/activity"
 	"getsturdy.com/api/pkg/analytics"
 	"getsturdy.com/api/pkg/auth"
 	"getsturdy.com/api/pkg/changes/message"
 	"getsturdy.com/api/pkg/codebases"
 	"getsturdy.com/api/pkg/events/v2"
 	"getsturdy.com/api/pkg/github"
-	"getsturdy.com/api/pkg/github/enterprise/client"
-	"getsturdy.com/api/pkg/github/enterprise/vcs"
+	"getsturdy.com/api/pkg/github/api"
+	github_client "getsturdy.com/api/pkg/github/enterprise/client"
+	vcs_github "getsturdy.com/api/pkg/github/enterprise/vcs"
 	gqlerrors "getsturdy.com/api/pkg/graphql/errors"
 	"getsturdy.com/api/pkg/users"
 	"getsturdy.com/api/pkg/workspaces"
+	db_workspaces "getsturdy.com/api/pkg/workspaces/db"
+	"getsturdy.com/api/vcs"
 )
 
 var ErrNotFound = errors.New("not found")
@@ -162,21 +166,197 @@ func (svc *Service) MergePullRequest(ctx context.Context, ws *workspaces.Workspa
 		return fmt.Errorf("pull request was not merged")
 	}
 
-	// This endpoint is not marking the PR as merged
-	// That happens when GitHub is sending the webhook to us
-	// Doing so, allows us to handle PR merges both from our and github's UI in the same way
-	//
-	// TODO: do not wait for the webhook, we already know everything we need
+	apiPR, _, err := userAuthClient.PullRequests.Get(ctx, ghInstallation.Owner, ghRepo.Name, pr.GitHubPRNumber)
+	if err != nil {
+		return fmt.Errorf("failed to get pull request: %w", err)
+	}
+
+	return svc.UpdatePullRequestFromGitHub(ctx, ghRepo, pr, api.ConvertPullRequest(apiPR))
+}
+
+func getPRStatus(apiPR *api.PullRequest) github.PullRequestState {
+	switch apiPR.GetState() {
+	case "open":
+		return github.PullRequestStateOpen
+	case "closed":
+		if apiPR.GetMerged() {
+			return github.PullRequestStateMerged
+		}
+		return github.PullRequestStateClosed
+	default:
+		return github.PullRequestStateUnknown
+	}
+}
+
+func (svc *Service) accessToken(ctx context.Context, repo *github.Repository) (string, error) {
+	installation, err := svc.gitHubInstallationRepo.GetByInstallationID(repo.InstallationID)
+	if err != nil {
+		return "", fmt.Errorf("could not get installation: %w", err)
+	}
+
+	accessToken, err := github_client.GetAccessToken(ctx, svc.logger, svc.gitHubAppConfig, installation, repo.GitHubRepositoryID, svc.gitHubRepositoryRepo, svc.gitHubInstallationClientProvider)
+	if err != nil {
+		return "", fmt.Errorf("failed to get access token: %w", err)
+	}
+
+	return accessToken, nil
+}
+
+func (svc *Service) UpdatePullRequestFromGitHub(
+	ctx context.Context,
+	repo *github.Repository,
+	pr *github.PullRequest,
+	gitHubPR *api.PullRequest,
+) error {
+	now := time.Now()
+	pr.UpdatedAt = &now
+	pr.ClosedAt = gitHubPR.ClosedAt
+	pr.MergedAt = gitHubPR.MergedAt
+	pr.State = getPRStatus(gitHubPR)
+	if err := svc.gitHubPullRequestRepo.Update(ctx, pr); err != nil {
+		return fmt.Errorf("failed to update github pull request in db: %w", err)
+	}
+
+	defer func() {
+		// make sure we send pr updated event after this function returns in any case
+		if err := svc.eventsPublisher.GitHubPRUpdated(ctx, events.Codebase(pr.CodebaseID), pr); err != nil {
+			svc.logger.Error("failed to send codebase event", zap.Error(err))
+			// do not fail
+		}
+	}()
+
+	accessToken, err := svc.accessToken(ctx, repo)
+	if err != nil {
+		return fmt.Errorf("failed to get github access token: %w", err)
+	}
+
+	ws, err := svc.workspaceReader.Get(pr.WorkspaceID)
+	if errors.Is(err, sql.ErrNoRows) {
+		svc.logger.Warn("handled a github pull request webhook for non-existing workspace",
+			zap.String("workspace_id", pr.WorkspaceID),
+			zap.String("github_pr_id", pr.ID),
+			zap.String("github_pr_link", gitHubPR.GetHTMLURL()),
+		)
+		return nil // noop
+	} else if err != nil {
+		return fmt.Errorf("failed to get workspace from db: %w", err)
+	}
+
+	// make context user context to connect all events to the same user
+	ctx = auth.NewUserContext(ctx, ws.UserID)
+
+	// update ws
+	if pr.Importing {
+		newDescription, err := DescriptionFromPullRequest(gitHubPR)
+		if err != nil {
+			return fmt.Errorf("failed to build description: %w", err)
+		}
+		ws.DraftDescription = newDescription
+		if err := svc.workspaceWriter.UpdateFields(ctx, ws.ID, db_workspaces.SetDraftDescription(newDescription)); err != nil {
+			return fmt.Errorf("failed to update workspace: %w", err)
+		}
+	}
+
+	if shouldArchive := pr.State == github.PullRequestStateClosed && pr.Importing; shouldArchive {
+		// if pr is closed and importing, archive it
+		return svc.workspacesService.Archive(ctx, ws)
+	}
+
+	if shouldUnarchive := pr.State == github.PullRequestStateOpen && pr.Importing; shouldUnarchive {
+		// if pr is open and not importing, unarchive it
+		if err := svc.workspacesService.Unarchive(ctx, ws); err != nil {
+			return fmt.Errorf("failed to unarchive workspace: %w", err)
+		}
+	}
+
+	if shouldUpdate := pr.State != github.PullRequestStateMerged; shouldUpdate {
+		// if the PR is not merged, fetch the latest PR data from GitHub
+		return svc.UpdatePullRequest(ctx, pr, gitHubPR, accessToken, ws)
+	}
+
+	// merge PR
+
+	// pull from github if sturdy doesn't have the commits
+	if err := svc.pullFromGitHubIfCommitNotExists(pr.CodebaseID, []string{
+		gitHubPR.GetMergeCommitSHA(),
+		gitHubPR.GetBase().GetSHA(),
+	}, accessToken, repo.TrackedBranch); err != nil {
+		return fmt.Errorf("failed to pullFromGitHubIfCommitNotExists: %w", err)
+	}
+
+	ch, err := svc.changeService.CreateWithCommitAsParent(ctx, ws, gitHubPR.GetMergeCommitSHA(), gitHubPR.GetBase().GetSHA())
+	if err != nil {
+		return fmt.Errorf("failed to create change: %w", err)
+	}
+
+	svc.analyticsService.Capture(ctx, "pull request merged",
+		analytics.DistinctID(ws.UserID.String()),
+		analytics.CodebaseID(ws.CodebaseID),
+		analytics.Property("workspace_id", ws.ID),
+		analytics.Property("github", true),
+		analytics.Property("importing", pr.Importing),
+	)
+
+	// send change to ci
+	if err := svc.buildQueue.EnqueueChange(ctx, ch); err != nil {
+		svc.logger.Error("failed to enqueue change", zap.Error(err))
+		// do not fail
+	}
+
+	// all workspaces that were up to date with trunk are not up to data anymore
+	if err := svc.workspaceWriter.UnsetUpToDateWithTrunkForAllInCodebase(ws.CodebaseID); err != nil {
+		return fmt.Errorf("failed to unset up to date with trunk for all in codebase: %w", err)
+	}
+
+	// Create workspace activity that it has created a change
+	if err := svc.activitySender.Codebase(ctx, ws.CodebaseID, ws.ID, ws.UserID, activity.TypeCreatedChange, string(ch.ID)); err != nil {
+		return fmt.Errorf("failed to create workspace activity: %w", err)
+	}
+
+	// copy all workspace activities to change activities
+	if err := svc.activityService.SetChange(ctx, ws.ID, ch.ID); err != nil {
+		return fmt.Errorf("failed to set change: %w", err)
+	}
+
+	if err := svc.commentsService.MoveCommentsFromWorkspaceToChange(ctx, ws.ID, ch.ID); err != nil {
+		return fmt.Errorf("failed to migrate comments: %w", err)
+	}
+
+	// Archive the workspace
+	if err := svc.workspacesService.ArchiveWithChange(ctx, ws, ch); err != nil {
+		return fmt.Errorf("failed to archive workspace: %w", err)
+	}
+
 	return nil
 }
 
-func Filter[T any](in []T, cond func(T) bool) (out []T) {
-	for _, v := range in {
-		if cond(v) {
-			out = append(out, v)
-		}
+func (svc *Service) pullFromGitHubIfCommitNotExists(codebaseID codebases.ID, commitShas []string, accessToken, trackedBranchName string) error {
+	shouldPull := false
+
+	if err := svc.executorProvider.New().
+		GitRead(func(repo vcs.RepoGitReader) error {
+			for _, sha := range commitShas {
+				if _, err := repo.GetCommitDetails(sha); err != nil {
+					shouldPull = true
+				}
+			}
+			return nil
+		}).
+		ExecTrunk(codebaseID, "pullFromGitHubIfCommitNotExists.Check"); err != nil {
+		return fmt.Errorf("failed to fetch changes from github: %w", err)
 	}
-	return out
+
+	if !shouldPull {
+		return nil
+	}
+
+	if err := svc.executorProvider.New().
+		GitWrite(vcs_github.FetchTrackedToSturdytrunk(accessToken, "refs/heads/"+trackedBranchName)).
+		ExecTrunk(codebaseID, "pullFromGitHubIfCommitNotExists.Pull"); err != nil {
+		return fmt.Errorf("failed to fetch changes from github: %w", err)
+	}
+
+	return nil
 }
 
 func (svc *Service) GetPullRequestForWorkspace(workspaceID string) (*github.PullRequest, error) {
@@ -291,7 +471,7 @@ func (svc *Service) CreateOrUpdatePullRequest(ctx context.Context, user *users.U
 		return nil, fmt.Errorf("failed to prepare branch: %w", err)
 	}
 
-	accessToken, err := client.GetAccessToken(ctx, logger, svc.gitHubAppConfig, ghInstallation, ghRepo.GitHubRepositoryID, svc.gitHubRepositoryRepo, svc.gitHubInstallationClientProvider)
+	accessToken, err := github_client.GetAccessToken(ctx, logger, svc.gitHubAppConfig, ghInstallation, ghRepo.GitHubRepositoryID, svc.gitHubRepositoryRepo, svc.gitHubInstallationClientProvider)
 	if err != nil {
 		return nil, err
 	}
@@ -308,9 +488,9 @@ func (svc *Service) CreateOrUpdatePullRequest(ctx context.Context, user *users.U
 	// If the repository is _empty_ there is a risk that the branch pushed for the PR is the first branch pushed to GH
 	// If this is the case, first push the sturdytrunk to be the new "master"/"main".
 	// This is done _without_ force, to not screw anything up if we're in the wrong.
-	if err := vcs.HaveTrackedBranch(svc.executorProvider, ws.CodebaseID, ghRepo.TrackedBranch); err != nil {
+	if err := vcs_github.HaveTrackedBranch(svc.executorProvider, ws.CodebaseID, ghRepo.TrackedBranch); err != nil {
 		logger.Info("pushing sturdytrunk to github")
-		userVisibleError, pushTrunkErr := vcs.PushBranchToGithubSafely(svc.executorProvider, ws.CodebaseID, "sturdytrunk", ghRepo.TrackedBranch, accessToken)
+		userVisibleError, pushTrunkErr := vcs_github.PushBranchToGithubSafely(svc.executorProvider, ws.CodebaseID, "sturdytrunk", ghRepo.TrackedBranch, accessToken)
 		if pushTrunkErr != nil {
 			logger.Error("failed to push trunk to github (github is source of truth)", zap.Error(pushTrunkErr))
 
@@ -327,7 +507,7 @@ func (svc *Service) CreateOrUpdatePullRequest(ctx context.Context, user *users.U
 		logger.Info("github have a default branch, not pushing sturdytrunk")
 	}
 
-	userVisibleError, pushErr := vcs.PushBranchToGithubWithForce(svc.executorProvider, ws.CodebaseID, prBranch, remoteBranchName, *ghUser.AccessToken)
+	userVisibleError, pushErr := vcs_github.PushBranchToGithubWithForce(svc.executorProvider, ws.CodebaseID, prBranch, remoteBranchName, *ghUser.AccessToken)
 	if pushErr != nil {
 		logger.Error("failed to push to github (github is source of truth)", zap.Error(pushErr))
